Added major improvements to the controls and gameplay
Tetrominos are no longer distributed in a random orientation. Improved the random tetrominos generation algorithm, which now tries to give the player the tetrominos it has received less. Improved the tetrominos rotation array to be less annoying with tetrominos who have only 2 rotations. Tetrominos no longer hit the ceiling when rotated right as they spawn. Improved automatic moving of Tetrominos when the player press and hold a key. When rotated, the game will try to find a place where the tetromino doesn't hit anything, for example, the sides of the grid.