appdata: Remove documentation license
The code in this project is covered by GPL-2.0+. The documentation is
covered by CC-BY-SA-3.0. As outlined on the GNOME Software wiki,
combining these two licenses with 'and' produces incorrect results, with
GNOME Software considering the project as a whole to be non-free (since
OSI does not consider CC-BY-SA-3.0 to be a free license for software).

Instead, list only the code's license in the appdata. This is
recommended by the appstream specification, and causes Software to
correctly label this application as free software.

Fixes: https://gitlab.gnome.org/GNOME/quadrapassel/-/issues/43

https://gitlab.gnome.org/GNOME/gnome-software/-/wikis/software-metadata#user-content-how-to-fix-incorrect-licensing-information
https://www.freedesktop.org/software/appstream/docs/chap-Metadata.html#tag-project_license