scores: fixes and improvements
This fixes a bug where the game's score would not be
recorded when the user quit the app.
This also removes the limit on how many scores are shown in
the scores dialog, and makes scores achieved with difficult
blocks turned on go into a '-Difficult' category. This adds
extra_info to new scores, this extra_info contains the rows
destroyed and the level.