menu_button: make screen readers work with the main menu
In most cases, only focused text is read by screen readers,
but we want the game area to have focus (so that it can receive
input). The fix is to add a mechanism to unfocus the main menu
when it is not in use, allowing the user to use the main menu
and not have to refocus the game area (by pressing tab), which
is annoying.